Factor out useFileUploads into a reusable hook
